Latest sync up

+++ sbe-benchmarks/src/main/cpp/CodecBench.h
@@ -1,11 +1,11 @@
 /*
- * Copyright 2013-2018 Real Logic Ltd.
+ * Copyright 2013-2021 Real Logic Limited.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ * https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -23,12 +23,12 @@ class CodecBench
 public:
     int encode_buffer(char *buffer, const int bufferLength)
     {
-        return static_cast<Derived *>(this)->encode(buffer, bufferLength);
+        return static_cast<int>(static_cast<Derived *>(this)->encode(buffer, bufferLength));
     };
 
     int decode_buffer(const char *buffer, const int bufferLength)
     {
-        return static_cast<Derived *>(this)->decode(buffer, bufferLength);
+        return static_cast<int>(static_cast<Derived *>(this)->decode(buffer, bufferLength));
     };
 
     /*
@@ -97,7 +97,9 @@ public:
         {
             ptr += encode_buffer(ptr, bufferLength);
         }
+
         ptr = buffer;
+
         for (int i = 0; i < n; i++)
         {
             ptr += decode_buffer(ptr, bufferLength);
